# storage: use `SetOptions()` when reusing Pebble iterators