@erikgrinaker 2022-05-23
This should be good to go now, PTAL.

A quick `kv95` run shows a slight performance regression here. I think we can merge this anyway, and work on performance optimizations later:

```
master:

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  180.0s        0       19341212       107450.5      1.6      1.1      4.5     10.5    159.4  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  180.0s        0        1017946         5655.2      4.0      3.3      8.9     15.2    121.6  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  180.0s        0       20359158       113105.7      1.7      1.2      5.0     11.0    159.4  

setoptions:

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  180.0s        0       18849240       104717.1      1.6      1.2      4.7     10.5    117.4  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  180.0s        0         992337         5512.9      4.1      3.4      9.4     15.7    201.3  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  180.0s        0       19841577       110230.1      1.7      1.2      5.2     11.0    201.3  
```